+++ extension/src/ALObject/ALCodeLine.ts
@@ -1,3 +1,5 @@
+import { ignoreCodeLinePattern } from "../constants";
+
 export class ALCodeLine {
   lineNo: number;
   code: string;
@@ -9,5 +11,42 @@ export class ALCodeLine {
     if (indentation) {
       this.indentation = indentation;
     }
+  }
+  static fromString(code: string): ALCodeLine[] {
+    const alCodeLines: ALCodeLine[] = [];
+    let lineNo = 0;
+
+    code
+      .replace(/(\r\n|\n)/gm, "\n")
+      .split("\n")
+      .forEach((line) => {
+        alCodeLines.push(new ALCodeLine(line, lineNo));
+        lineNo++;
+      });
+
+    return alCodeLines;
+  }
+
+  public isInsignificant(): boolean {
+    // Comments, compiler directives and whitespace
+    const ignoreRegex = new RegExp(ignoreCodeLinePattern, "im");
+    const ignoreMatch = this.code.match(ignoreRegex);
+    return null !== ignoreMatch;
+  }
+
+  public isWhitespace(): boolean {
+    return this.code.trim() === "";
+  }
+
+  public isXmlComment(): boolean {
+    return null !== this.code.match(/^\s*\/\/\/.*/i);
+  }
+
+  public isCompilerDirective(): boolean {
+    return null !== this.code.match(/^\s*#.*/i);
+  }
+
+  public matchesPattern(regexp: string | RegExp): boolean {
+    return null !== this.code.match(regexp);
   }
 }