+++ extension/CHANGELOG.md
@@ -11,6 +11,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Changes:
   - Dependency on module jDataView removed and replaced with internal implementation of a BinaryReader.
+  - Improved handling of how procedures are identified with better handling of compiler directives, XmlComments and comments.
 
 ## [1.2.0] - 2021-06-03